Move system commands and events into sub-packages
