+++ server/src/main/java/io/spine/server/commandbus/CommandReceivedTap.java
@@ -25,7 +25,7 @@ import io.spine.core.Command;
 import io.spine.core.TenantId;
 import io.spine.server.bus.BusFilter;
 import io.spine.server.type.CommandEnvelope;
-import io.spine.system.server.CommandReceived;
+import io.spine.system.server.event.CommandReceived;
 import io.spine.system.server.SystemWriteSide;
 import io.spine.system.server.WriteSideFunction;
