+++ server/src/main/java/io/spine/server/commandbus/Split.java
@@ -25,7 +25,7 @@ import io.spine.annotation.Internal;
 import io.spine.base.CommandMessage;
 import io.spine.core.Command;
 import io.spine.server.type.CommandEnvelope;
-import io.spine.system.server.CommandSplit;
+import io.spine.system.server.event.CommandSplit;
 import io.spine.system.server.SystemWriteSide;
 
 import static com.google.common.base.Preconditions.checkState;
