+++ server/src/main/proto/spine/system/server/event_lifecycle_commands.proto
@@ -27,7 +27,7 @@ import "spine/options.proto";
 option (type_url_prefix) = "type.spine.io";
 option (internal_all) = true;
 
-option java_package = "io.spine.system.server";
+option java_package = "io.spine.system.server.command";
 option java_outer_classname = "EventLifecycleCommandsProto";
 option java_multiple_files = true;
 
@@ -38,18 +38,18 @@ import "spine/core/event.proto";
 message MarkCausedCommand {
 
     // The ID of the event.
-    spine.core.EventId id = 1;
+    core.EventId id = 1;
 
     // The ID of the command generated in response to the event.
-    spine.core.CommandId produced = 2 [(required) = true];
+    core.CommandId produced = 2 [(required) = true];
 }
 
 // Marks a domain event as one which caused two or more commands.
 message MarkCausedCommands {
 
     // The ID of the event.
-    spine.core.EventId id = 1;
+    core.EventId id = 1;
 
     // Two or more IDs of commands generated in response to the event.
-    repeated spine.core.CommandId produced = 2 [(required) = true];
+    repeated core.CommandId produced = 2 [(required) = true];
 }
