+++ server/src/main/proto/spine/system/server/event_lifecycle_events.proto
@@ -27,7 +27,7 @@ import "spine/options.proto";
 option (type_url_prefix) = "type.spine.io";
 option (internal_all) = true;
 
-option java_package = "io.spine.system.server";
+option java_package = "io.spine.system.server.event";
 option java_outer_classname = "EventLifecycleEventsProto";
 option java_multiple_files = true;
