+++ server/src/main/proto/spine/system/server/history_events.proto
@@ -27,7 +27,7 @@ import "spine/options.proto";
 option (type_url_prefix) = "type.spine.io";
 option (internal_all) = true;
 
-option java_package = "io.spine.system.server";
+option java_package = "io.spine.system.server.event";
 option java_outer_classname = "HistoryEventsProto";
 option java_multiple_files = true;
 
@@ -117,7 +117,7 @@ message EventImported {
     EntityHistoryId receiver = 1 [(required) = true];
 
     // The info about the event.
-    spine.core.EventId event_id = 2 [(required) = true, (valid) = true];
+    core.EventId event_id = 2 [(required) = true, (valid) = true];
 
     // The time when the event was imported.
     google.protobuf.Timestamp when_imported = 3 [(required) = true, (valid) = true];
