+++ server/src/main/proto/spine/system/server/history_rejections.proto
@@ -27,7 +27,7 @@ import "spine/options.proto";
 option (type_url_prefix) = "type.spine.io";
 option (internal_all) = true;
 
-option java_package = "io.spine.system.server";
+option java_package = "io.spine.system.server.rejection";
 option java_outer_classname = "HistoryRejections";
 
 import "google/protobuf/timestamp.proto";
