+++ server/src/test/java/io/spine/server/commandbus/CommandAckMonitorTest.java
@@ -36,8 +36,8 @@ import io.spine.server.bus.Buses;
 import io.spine.server.entity.rejection.CannotModifyArchivedEntity;
 import io.spine.server.event.RejectionEnvelope;
 import io.spine.server.type.CommandEnvelope;
-import io.spine.system.server.CommandAcknowledged;
-import io.spine.system.server.CommandErrored;
+import io.spine.system.server.event.CommandAcknowledged;
+import io.spine.system.server.event.CommandErrored;
 import io.spine.system.server.MemoizingWriteSide;
 import io.spine.system.server.NoOpSystemWriteSide;
 import io.spine.test.commandbus.CmdBusStartProject;
