+++ server/src/test/java/io/spine/server/entity/EntityLifecycleTest.java
@@ -27,8 +27,8 @@ import com.google.protobuf.Timestamp;
 import io.spine.base.Identifier;
 import io.spine.base.Time;
 import io.spine.core.EventId;
-import io.spine.system.server.EntityCreated;
-import io.spine.system.server.EntityStateChanged;
+import io.spine.system.server.event.EntityCreated;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.system.server.MemoizedSystemMessage;
 import io.spine.system.server.MemoizingWriteSide;
 import io.spine.system.server.NoOpSystemWriteSide;
