+++ server/src/test/java/io/spine/server/model/declare/given/MethodParamsTestEnv.java
@@ -29,7 +29,7 @@ import io.spine.core.CommandContext;
 import io.spine.core.UserId;
 import io.spine.server.model.declare.ParameterSpec;
 import io.spine.server.type.CommandEnvelope;
-import io.spine.system.server.ScheduleCommand;
+import io.spine.system.server.command.ScheduleCommand;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
