+++ server/src/test/java/io/spine/server/procman/PmSystemEventWatcherTest.java
@@ -33,11 +33,11 @@ import io.spine.server.event.DuplicateEventException;
 import io.spine.server.procman.given.delivery.GivenMessage;
 import io.spine.server.type.CommandEnvelope;
 import io.spine.server.type.EventEnvelope;
-import io.spine.system.server.CommandDispatchedToHandler;
-import io.spine.system.server.CommandDispatchedToHandlerVBuilder;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.EventDispatchedToReactor;
-import io.spine.system.server.EventDispatchedToReactorVBuilder;
+import io.spine.system.server.event.CommandDispatchedToHandler;
+import io.spine.system.server.event.CommandDispatchedToHandlerVBuilder;
+import io.spine.system.server.event.EventDispatchedToReactor;
+import io.spine.system.server.event.EventDispatchedToReactorVBuilder;
 import io.spine.testing.server.TestEventFactory;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.BeforeEach;
@@ -49,8 +49,8 @@ import static com.google.common.testing.NullPointerTester.Visibility.PACKAGE;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.base.Identifier.pack;
 import static io.spine.base.Time.currentTime;
-import static io.spine.system.server.HistoryRejections.CannotDispatchCommandTwice;
-import static io.spine.system.server.HistoryRejections.CannotDispatchEventTwice;
+import static io.spine.system.server.rejection.HistoryRejections.CannotDispatchCommandTwice;
+import static io.spine.system.server.rejection.HistoryRejections.CannotDispatchEventTwice;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
