+++ server/src/test/java/io/spine/server/projection/ProjectionShould.java
@@ -43,7 +43,7 @@ import io.spine.string.StringifierRegistry;
 import io.spine.string.Stringifiers;
 import io.spine.system.server.DispatchedMessageId;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.EntityStateChanged;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.test.projection.Project;
 import io.spine.test.projection.ProjectId;
 import io.spine.test.projection.ProjectTaskNames;
