+++ server/src/test/java/io/spine/server/projection/ProjectionSystemEventWatcherTest.java
@@ -30,8 +30,8 @@ import io.spine.server.event.DuplicateEventException;
 import io.spine.server.type.EventEnvelope;
 import io.spine.server.type.given.GivenEvent;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.EventDispatchedToSubscriber;
-import io.spine.system.server.EventDispatchedToSubscriberVBuilder;
+import io.spine.system.server.event.EventDispatchedToSubscriber;
+import io.spine.system.server.event.EventDispatchedToSubscriberVBuilder;
 import io.spine.testing.server.TestEventFactory;
 import io.spine.type.TypeUrl;
 import org.junit.jupiter.api.BeforeEach;
@@ -42,7 +42,7 @@ import org.junit.jupiter.api.Test;
 import static com.google.common.testing.NullPointerTester.Visibility.PACKAGE;
 import static io.spine.base.Identifier.newUuid;
 import static io.spine.base.Time.currentTime;
-import static io.spine.system.server.HistoryRejections.CannotDispatchEventTwice;
+import static io.spine.system.server.rejection.HistoryRejections.CannotDispatchEventTwice;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.ArgumentMatchers.any;
