+++ server/src/test/java/io/spine/system/server/CommandLifecycleTest.java
@@ -38,6 +38,13 @@ import io.spine.core.UserId;
 import io.spine.server.BoundedContext;
 import io.spine.server.DefaultRepository;
 import io.spine.server.commandbus.CommandBus;
+import io.spine.system.server.event.CommandAcknowledged;
+import io.spine.system.server.event.CommandDispatched;
+import io.spine.system.server.event.CommandErrored;
+import io.spine.system.server.event.CommandHandled;
+import io.spine.system.server.event.CommandReceived;
+import io.spine.system.server.event.CommandRejected;
+import io.spine.system.server.event.TargetAssignedToCommand;
 import io.spine.system.server.given.command.CommandLifecycleWatcher;
 import io.spine.system.server.given.command.CompanyAggregate;
 import io.spine.system.server.given.command.CompanyNameProcman;
