+++ server/src/test/java/io/spine/system/server/EntityHistoryTest.java
@@ -34,6 +34,15 @@ import io.spine.option.EntityOption;
 import io.spine.people.PersonName;
 import io.spine.server.BoundedContext;
 import io.spine.server.DefaultRepository;
+import io.spine.system.server.event.CommandDispatchedToHandler;
+import io.spine.system.server.event.EntityArchived;
+import io.spine.system.server.event.EntityCreated;
+import io.spine.system.server.event.EntityDeleted;
+import io.spine.system.server.event.EntityExtractedFromArchive;
+import io.spine.system.server.event.EntityRestored;
+import io.spine.system.server.event.EntityStateChanged;
+import io.spine.system.server.event.EventDispatchedToReactor;
+import io.spine.system.server.event.EventDispatchedToSubscriber;
 import io.spine.system.server.given.entity.HistoryEventWatcher;
 import io.spine.system.server.given.entity.PersonAggregate;
 import io.spine.system.server.given.entity.PersonNamePart;
