+++ server/src/test/java/io/spine/system/server/MirrorEventsTest.java
@@ -23,6 +23,11 @@ package io.spine.system.server;
 import com.google.protobuf.Empty;
 import io.spine.base.EventMessage;
 import io.spine.core.Event;
+import io.spine.system.server.event.EntityArchived;
+import io.spine.system.server.event.EntityDeleted;
+import io.spine.system.server.event.EntityExtractedFromArchive;
+import io.spine.system.server.event.EntityRestored;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.testing.server.TestEventFactory;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
