+++ server/src/test/java/io/spine/system/server/given/mirror/ProjectionTestEnv.java
@@ -25,12 +25,12 @@ import io.spine.client.EntityId;
 import io.spine.core.EventId;
 import io.spine.core.Version;
 import io.spine.system.server.DispatchedMessageId;
-import io.spine.system.server.EntityArchived;
-import io.spine.system.server.EntityDeleted;
-import io.spine.system.server.EntityExtractedFromArchive;
+import io.spine.system.server.event.EntityArchived;
+import io.spine.system.server.event.EntityDeleted;
+import io.spine.system.server.event.EntityExtractedFromArchive;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.EntityRestored;
-import io.spine.system.server.EntityStateChanged;
+import io.spine.system.server.event.EntityRestored;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.system.server.MirrorId;
 import io.spine.test.system.server.Video;
 import io.spine.type.TypeUrl;
