+++ server/src/test/java/io/spine/system/server/given/mirror/RepositoryTestEnv.java
@@ -29,10 +29,10 @@ import io.spine.core.Event;
 import io.spine.core.EventId;
 import io.spine.net.Url;
 import io.spine.system.server.DispatchedMessageId;
-import io.spine.system.server.EntityArchived;
-import io.spine.system.server.EntityDeleted;
+import io.spine.system.server.event.EntityArchived;
+import io.spine.system.server.event.EntityDeleted;
 import io.spine.system.server.EntityHistoryId;
-import io.spine.system.server.EntityStateChanged;
+import io.spine.system.server.event.EntityStateChanged;
 import io.spine.test.system.server.Photo;
 import io.spine.test.system.server.PhotoId;
 import io.spine.test.system.server.PhotoVBuilder;
